backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
django==5.2.4; python_version >= '3.10'
djangorestframework==3.16.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
